# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "私募基金监控平台 API",
    "version": "1.0.0",
    "description": "私募排排网基金数据抓取与监控平台，支持 OCR 净值识别、多标签筛选、数据导出。"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000",
      "description": "本地开发环境"
    }
  ],
...
  "paths": {
...
    "/api/crawl": {
      "post": {
        "tags": ["爬虫"],
        "summary": "统一爬取接口",
        "description": "支持同步/异步执行，单个/多个标签爬取。sync=true 时阻塞返回完整结果；sync=false 时异步立即返回。",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tags": {
                    "type": "array",
                    "items": { "type": "string" },
                    "default": ["private"],
                    "example": ["private", "public"],
                    "description": "标签列表，不传则爬取全部标签"
                  },
                  "sync": {
                    "type": "boolean",
                    "default": false,
                    "description": "true=同步阻塞返回完整结果，false=异步立即返回"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "同步爬取结果或异步已启动",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CrawlResponse" }
              }
            }
          },
          "409": {
            "description": "爬虫正在运行中",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "running" },
                    "message": { "type": "string" },
                    "last_crawl": { "type": "string", "nullable": true }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/crawl/records": {
      "get": {
        "tags": ["爬虫"],
        "summary": "获取爬取记录列表",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "tag", "in": "query", "schema": { "type": "string" }, "description": "按标签筛选" }
        ],
        "responses": {
          "200": {
            "description": "爬取记录分页列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "items": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/CrawlRecord" }
                    },
                    "total": { "type": "integer" },
                    "page": { "type": "integer" },
                    "per_page": { "type": "integer" },
                    "pages": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/funds": {
      "get": {
        "tags": ["基金数据"],
        "summary": "查询基金列表（分页 + 多维筛选）",
        "description": "从数据库查询，支持标签、日期、基金名称、基金代码筛选",
        "parameters": [
          { "name": "tag", "in": "query", "schema": { "type": "string", "default": "private" }, "description": "基金标签" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "name", "in": "query", "schema": { "type": "string" }, "description": "基金名称模糊搜索" },
          { "name": "code", "in": "query", "schema": { "type": "string" }, "description": "基金代码精确匹配" },
          { "name": "date", "in": "query", "schema": { "type": "string" }, "description": "爬取日期（YYYY-MM-DD）" },
          { "name": "sort", "in": "query", "schema": { "type": "string", "default": "crawl_time" }, "description": "排序字段" },
          { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" } }
        ],
        "responses": {
          "200": {
            "description": "基金分页列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tag": { "type": "string" },
                    "items": { "type": "array", "items": { "$ref": "#/components/schemas/Fund" } },
                    "total": { "type": "integer" },
                    "page": { "type": "integer" },
                    "per_page": { "type": "integer" },
                    "pages": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/funds/latest": {
      "get": {
        "tags": ["基金数据"],
        "summary": "获取最新基金数据（每个基金取最新一条）",
        "parameters": [
          { "name": "tag", "in": "query", "schema": { "type": "string", "default": "private" } }
        ],
        "responses": {
          "200": {
            "description": "最新基金列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tag": { "type": "string" },
                    "items": { "type": "array", "items": { "$ref": "#/components/schemas/Fund" } },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/fund/{fund_name}": {
      "get": {
        "tags": ["基金数据"],
        "summary": "获取单个基金的详细历史数据",
        "parameters": [
          { "name": "fund_name", "in": "path", "required": true, "schema": { "type": "string" }, "description": "基金名称（URL 编码）" },
          { "name": "tag", "in": "query", "schema": { "type": "string", "default": "private" } }
        ],
        "responses": {
          "200": {
            "description": "基金历史记录",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tag": { "type": "string" },
                    "fund_name": { "type": "string" },
                    "records": { "type": "array", "items": { "$ref": "#/components/schemas/Fund" } },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          },
          "404": { "description": "基金不存在" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Fund": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "tag": { "type": "string", "description": "标签：private/public/money" },
          "fund_name": { "type": "string", "description": "基金名称" },
          "fund_code": { "type": "string", "description": "基金代码" },
          "strategy": { "type": "string", "description": "投资策略" },
          "net_value_date": { "type": "string", "description": "净值日期" },
          "net_value": { "type": "string", "description": "最新净值（OCR识别）" },
          "net_change": { "type": "string", "description": "净值变动" },
          "this_year": { "type": "string", "description": "今年来收益" },
          "one_year": { "type": "string", "description": "近一年收益" },
          "three_year": { "type": "string", "description": "近三年收益" },
          "since_inception": { "type": "string", "description": "成立来收益" },
          "drawdown": { "type": "string", "description": "历史最大回撤" },
          "crawl_date": { "type": "string", "description": "爬取日期" },
          "crawl_time": { "type": "string", "description": "爬取时间" }
        }
      },
      "CrawlRecord": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "tag": { "type": "string" },
          "crawl_date": { "type": "string" },
          "start_time": { "type": "string" },
          "end_time": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["running", "success", "skipped", "failed"] },
          "total_funds": { "type": "integer", "nullable": true },
          "ocr_success": { "type": "integer", "nullable": true },
          "error_message": { "type": "string", "nullable": true }
        }
      },
      "CrawlResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "started" },
          "message": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "sync": { "type": "boolean" },
          "last_crawl": { "type": "string", "nullable": true }
        },
        "example_async": {
          "status": "started",
          "message": "爬取已在后台启动",
          "tags": ["private", "public"],
          "sync": false
        }
      },
...
    }
  }
}